NYT average logic
opens wordle bot in an active page. navigates to analysis and scrapes the NYT average for the day. using this method instead of fetch()ing and parsing DOM because NYT average is loaded with javascript :( this is getting overcomplicated, too hardcoded and takes so much tedious debugging of silent errors.